{"schema":"decodex/commit/1","summary":"Modularize orchestrator control-plane entrypoints","authority":"XY-1115"}
